updare evaluation

+++ code/train_nb.py
@@ -24,11 +24,11 @@ def train_naive_bayes():
     #     print(len(des_labels))
 
     # import IPython; IPython.embed()
-    X_train= vec_web_sites[:(vec_web_sites.shape[0])*0.95]
-    X_valid= vec_web_sites[(vec_web_sites.shape[0])*0.95:]
+    X_train= vec_web_sites[:int((vec_web_sites.shape[0])*0.95)]
+    X_valid= vec_web_sites[int((vec_web_sites.shape[0])*0.95):]
 
-    Y_train = labels[:(vec_web_sites.shape[0])*0.95]
-    Y_valid = labels[(vec_web_sites.shape[0])*0.95:]
+    Y_train = labels[:int(vec_web_sites.shape[0]*0.95)]
+    Y_valid = labels[int(vec_web_sites.shape[0]*0.95):]
 
 
     clf = gnb.fit(vec_des_data, des_labels)
